Added Log file name and logic to check if a file is already processed
Logic to check if a file is already parsed is now commented though.
This will read a file containing list of all the files that has been
already processed. A list of filenames is prepared, and the subsequent
files that are to be processed are checked to see if its already
existing in the list.